added Emitter4.jl
The implementation resides in its own module scope (Emitters).

NOTE: Although the emitter is part of the optical part of the package, its INCLUDE statement is not part of the includes in Optical.jl because we need to include it after the Visualization.jl file to override the visualization functions. This will be solved in the future once we have a better visualization implementation.