added a new implementation of the Axis Symmetric Optimization + UI
now supporting the optimization variable syntax of " optional lower bound <
VarName=StartValue < optional upper bound " (ex: 10<@=20<30 - @ will create an
automatic name for the variable).

lots of changes in the UI part.

zoom in/zoom out in the blink window - Ctrl-+ / Ctrl-- / Ctrl-0 to reset
F12 - toggle dev tools

added support for interactive HTML - using the three_js implementation of
WGLMakie to allow real 3D and 2D interaction with plots in the browser.

added the ability to include local assets in the UI, such as images and css files.

added the ability to open/run associated/default program for specific files (for
instance, editing in excel or another CSV associated program - should be tested
on other OSs)

added the ability to run JS scripts on the shell (not just the browser process)