Added a new "Surface" type: ConvexPolygon which support planner simple convex polygons (#213)
* added two new libraries dependencies - one is part of the Julia basic libs, the other one is for fast processing of 2D polygons.
Added a new "Surface" child: ConvexPolygon which support planner simple convex polygons. These can be used to create non-symmetric paraxial lenses.
Moved a couple of utility functions form the Emitters module to the Geometry module where they belong. These include right, up, forward and origins for type Transform.
Also added some tests for the new ConvexPolygon type.

* added some comments and changed the type of the local points to SVector{2, T}.
Computed the inverse of the local frame once in the constructor instead of inside the ray intersection test - should be more efficient.

* Removed the GeometricalPredicates from the project. This package was used to test whether a 2D point lies inside a convex polygon but was using too many allocations during the test.
I implemented the actual test and removed all the allocations.

* ConvexPolygon is now back to using a Vector{SVector{2,T}} instead of an SVector{N SVector{2, T}} due to the ripple effects that addition of N in the system.
Currently, this change leads to ParaxielLens which hold a union of shapes iw now referencing the ALL the shape because if the variable nature of ConvexPolygon. Didn't notice a huge difference in speed but we need to be aware of it.

* removed the normal(surface, x, y) for planner shapes.
removed the unused T type declaration from some functions.

Co-authored-by: BrianGun <36491850+BrianGun@users.noreply.github.com>